requirements.txt
colorama==0.4.6
iniconfig==2.1.0
logging==0.4.9.6
numpy==2.3.2
packaging==25.0
pluggy==1.6.0
Pygments==2.19.2
pytest==8.4.1
PyYAML==6.0.2
